[FE] MOVE: 폴더명 변경  #39
