+++ src/frontend/src/components/LeftNavBar/index.css.ts
