+++ src/frontend/src/pages/FriendPage/index.tsx
